% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallEshotgun.R
\name{checkLibraries}
\alias{checkLibraries}
\title{Check for required Library and Python Packages}
\usage{
checkLibraries(method = "auto", conda = "auto")
}
\arguments{
\item{method}{method of installation}

\item{conda}{environment}
}
\description{
The function will download all required Python Packages that
are rquired for the e-shotgun to run properly.
}
\value{
No return value, called for side effects
}
\examples{
checkLibraries()
}
