% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{modifiedBranin}
\alias{modifiedBranin}
\title{modifiedBranin}
\usage{
modifiedBranin(x)
}
\arguments{
\item{x}{matrix with the initial points}
}
\value{
m a matrix
}
\description{
The Branin Function applied to a matrix
}
