% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{modifiedEgg}
\alias{modifiedEgg}
\title{modifiedEgg}
\usage{
modifiedEgg(x)
}
\arguments{
\item{x}{matrix with the inital points}
}
\value{
a matrix
}
\description{
The Eggholder Function applied to a matrix
}
