% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{modifiedschwef}
\alias{modifiedschwef}
\title{modifiedschwef}
\usage{
modifiedschwef(x)
}
\arguments{
\item{x}{matrix with the initial points}
}
\value{
a matrix
}
\description{
The schwefel Function applied to a matrix
}
