% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{runSampleOpt}
\alias{runSampleOpt}
\title{runSampleOpt}
\usage{
runSampleOpt(fn, budget = 100)
}
\arguments{
\item{fn}{function of the testproblem}

\item{budget}{budget for the run}
}
\value{
No return value, called for visual showcasing
}
\description{
Optimizing a function with the e-shotgun
}
